pyproject.toml
[tool.poetry]
name = "update-docs-pls"
version = "0.1.0"
description = ""
authors = ["Andrés Riquelme <andresfranco.rs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
actions-toolkit = "^0.1.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
